requirements.txt
# POS SYSTEM - Python Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0.post1

# Payments
stripe==7.12.0

# SMS (Africa's Talking)
africastalking==1.2.7

# Redis (Caching & Sessions)
redis==5.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.26.0

# PDF Generation (Receipts)
reportlab==4.0.8

# Logging
structlog==24.1.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
black==24.1.0
isort==5.13.2
flake8==7.0.0
